backend/chatbot/api/requirement.txt
et-xmlfile==1.1.0
openpyxl==3.1.2
langchain
pandas
django
djangorestframework
environs
langchain-experimental
langchain-openai
django-cors-headers

